+++ app/src/components/Filters/StringFilter.tsx
@@ -17,7 +17,7 @@ import Results, { ResultsContainer } from "../Common/Results";
 import { Button } from "../FieldsSidebar";
 import { PopoutSectionTitle, TabOption } from "../utils";
 import { LIST_LIMIT } from "./StringFieldFilter";
-import { ItemAction } from "../Actions/utils";
+import { ItemAction } from "../Actions/ItemAction";
 import socket from "../../shared/connection";
 import { packageMessage } from "../../utils/socket";
 import { useTheme } from "../../utils/hooks";
